Finalised split classes
