+++ .gitignore
@@ -1,4 +1,38 @@
-Data_Download/key.py
 node_modules
-Data_Download/Sports_Data/
 APICaller/key.py
+Sports_Data/Afl/Leagues.json
+Sports_Data/Afl/Seasons.json
+Sports_Data/Afl/Status.json
+Sports_Data/American-football/Leagues.json
+Sports_Data/American-football/Seasons.json
+Sports_Data/American-football/Status.json
+Sports_Data/Baseball/Leagues.json
+Sports_Data/Baseball/Seasons.json
+Sports_Data/Baseball/Status.json
+Sports_Data/Basketball/Leagues.json
+Sports_Data/Basketball/Seasons.json
+Sports_Data/Basketball/Status.json
+Sports_Data/Football/Leagues.json
+Sports_Data/Football/seasons.json
+Sports_Data/Football/Status.json
+Sports_Data/Formula-1/Leagues.json
+Sports_Data/Formula-1/Seasons.json
+Sports_Data/Formula-1/Status.json
+Sports_Data/Handball/Leagues.json
+Sports_Data/Handball/Seasons.json
+Sports_Data/Handball/Status.json
+Sports_Data/Hockey/Leagues.json
+Sports_Data/Hockey/Seasons.json
+Sports_Data/Hockey/Status.json
+Sports_Data/Mma/Leagues.json
+Sports_Data/Mma/Seasons.json
+Sports_Data/Mma/Status.json
+Sports_Data/Nba/Leagues.json
+Sports_Data/Nba/Seasons.json
+Sports_Data/Nba/Status.json
+Sports_Data/Rugby/Leagues.json
+Sports_Data/Rugby/Seasons.json
+Sports_Data/Rugby/Status.json
+Sports_Data/Volleyball/Leagues.json
+Sports_Data/Volleyball/Seasons.json
+Sports_Data/Volleyball/Status.json
