+++ Tools/FileManager/FileClass.py
@@ -1,17 +1,21 @@
 import pathlib, json
-from Tools.Sports.sportclass import Sport as Class
+from Tools.Sports.SportClass import Sport as SportClass
 
-from Tools.APICaller import caller
+from Tools.APICaller import APIClass
 
-class FileManager(Class):
+class FileManager(SportClass):
     def __init__(self, sport, version):
         super().__init__(sport, version)
+        self.sport_folder = "Sports_Data/"+str(self.sport).capitalize()
 
     def create_folder(self, folder):
         path = pathlib.Path(folder) #set folder filepath
         path.mkdir(parents=True, exist_ok=True) #create folder
         return(path)
-        
+
+    def get_folder(self):
+        return(self.sport_folder)
+
     def create_file(self, path, filename, result):
         fn = filename+".json" #set up json file name
         filepath = path+"/"+fn #set file filepath
@@ -19,17 +23,27 @@ class FileManager(Class):
             json.dump(result, f, ensure_ascii=False, indent=4) #create file
 
     def download_data(self, folder, call, qualifiers=""):
-        API = caller.APICall(self.sport, self.version)
+        API = APIClass.APICall(self.sport, self.version)
         response = API.API_request(call, qualifiers)
         if "/" in call:
             self.create_file(folder, str(call).capitalize().rsplit("/", 1)[1], response) #Set the filename to anything beyond the '/' in the call, then create data file
         else:
             self.create_file(folder, str(call).capitalize(), response) #Create data file within folder
 
+    def get_status(self):
+        self.download_data(self.sport_folder,"status")
+
+    def get_leagues(self):
+        self.download_data(self.sport_folder,"leagues")
+    
+    def get_seasons(self):
+        if self.sport == "football":
+            self.download_data(self.sport_folder,"leagues/seasons")
+        else:
+            self.download_data(self.sport_folder,"seasons")
+
     def set_up_sport(self):
-        API = caller.APICall(self.sport, self.version)
-        getfolder = str(API.get_folder())
-        self.create_folder(getfolder)
-        API.get_status()
-        API.get_leagues()
-        API.get_seasons()
+        self.create_folder(self.get_folder())
+        self.get_status()
+        self.get_leagues()
+        self.get_seasons()
